requirements.txt
chromadb>=0.5.23
faiss-cpu>=1.9.0.post1
fastapi==0.115.7
kaggle==1.6.17
langchain==0.3.15
langchain-chroma==0.2.0
langchain-community==0.3.15
langchain-core==0.3.31
langchain-experimental>=0.3.4
langchain-huggingface==0.1.2
langchain-ollama>=0.2.2
langchain-openai==0.3.2
langgraph>=0.2.67
loguru==0.7.3
numpy==1.26.4
ollama>=0.4.7
pandas==2.0.2
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
rank-bm25==0.2.2
streamlit>=1.41.1
transformers==4.46.3